move rapidsnark to submodule
