+++ cmd/proof/go.mod
@@ -3,9 +3,9 @@ module github.com/iden3/go-rapidsnark/cmd/proof
 go 1.18
 
 require (
-	github.com/iden3/go-rapidsnark v0.0.0-20220513033005-077fb4535566
+	github.com/iden3/go-rapidsnark/prover v0.0.0-20220513033005-077fb4535566
 )
 
 replace (
-	github.com/iden3/go-rapidsnark => ../..
+	github.com/iden3/go-rapidsnark/prover => ../../prover
 )
