--- go.mod
@@ -1,3 +1,0 @@
-module github.com/iden3/go-rapidsnark
-
-go 1.18
