+++ prover/build_dynamic.go
@@ -1,6 +1,6 @@
 //go:build dynamic
 
-package rapidsnark
+package prover
 
 // #cgo LDFLAGS: -lrapidsnark -lgmp -lstdc++
 import "C"
