+++ prover/build_linux_amd64.go
@@ -1,6 +1,6 @@
 //go:build !dynamic
 
-package rapidsnark
+package prover
 
 // #cgo CFLAGS: -DUSE_VENDORED_RAPIDSNARK
 // #cgo LDFLAGS: ${SRCDIR}/rapidsnark_vendor/librapidsnark-linux-amd64.a ${SRCDIR}/rapidsnark_vendor/libgmp-linux-amd64.a -lstdc++
