+++ prover/go.mod
@@ -1,0 +1,8 @@
+module github.com/iden3/go-rapidsnark/prover
+
+go 1.18
+
+replace github.com/iden3/go-rapidsnark/types => ../types
+
+require github.com/iden3/go-rapidsnark/types v0.0.1
+
