+++ prover/rapidsnark_vendor/libgmp-darwin-amd64.a
