+++ prover/rapidsnark_vendor/librapidsnark-darwin-amd64.a
