+++ prover/rapidsnark_vendor/librapidsnark-darwin-arm64.a
