+++ prover/rapidsnark_vendor/prover.h
