+++ types/go.mod
@@ -1,0 +1,3 @@
+module github.com/iden3/go-rapidsnark/types
+
+go 1.18
