+++ witness/go.mod
@@ -1,0 +1,14 @@
+module github.com/iden3/go-rapidsnark/witness
+
+go 1.18
+
+require (
+	github.com/stretchr/testify v1.7.1
+	github.com/wasmerio/wasmer-go v1.0.4
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
+)
